+++ compose/ui/ui/src/skikoMain/kotlin/androidx/compose/ui/node/SnapshotInvalidationTracker.kt
@@ -14,16 +14,75 @@
  * limitations under the License.
  */
 
-package androidx.compose.ui.scene
+package androidx.compose.ui.node
 
+import androidx.compose.runtime.snapshots.Snapshot
 import androidx.compose.ui.createSynchronizedObject
 import androidx.compose.ui.synchronized
+
+/**
+ * SnapshotCommandList is a class that manages commands and invalidations for snapshot-based recomposition.
+ * It allows postponing execution of commands and performing them in the future.
+ *
+ * @param invalidate a function that is called whenever an invalidation is requested
+ */
+internal class SnapshotInvalidationTracker(
+    private val invalidate: () -> Unit = {}
+) {
+    private val snapshotChanges = CommandList(invalidate)
+    private var needLayout = true
+    private var needDraw = true
+
+    val hasInvalidations: Boolean
+        get() = needLayout || needDraw || snapshotChanges.hasCommands
+
+    fun requestLayout() {
+        needLayout = true
+        invalidate()
+    }
+
+    fun onLayout() {
+        // Apply changes from recomposition phase to layout phase
+        sendAndPerformSnapshotChanges()
+
+        needLayout = false
+    }
+
+    fun requestDraw() {
+        needDraw = true
+        invalidate()
+    }
+
+    fun onDraw() {
+        // Apply changes from layout phase to draw phase
+        sendAndPerformSnapshotChanges()
+
+        needDraw = false
+    }
+
+    /**
+     * Creates an observer for monitoring changes in the snapshot of an owner.
+     *
+     * @return the observer for monitoring snapshot changes
+     */
+    fun snapshotObserver() = OwnerSnapshotObserver { command ->
+        snapshotChanges.add(command)
+    }
+
+    /**
+     * Sends any pending apply notifications and performs the changes they cause.
+     */
+    fun sendAndPerformSnapshotChanges() {
+        Snapshot.sendApplyNotifications()
+        snapshotChanges.perform()
+    }
+}
 
 /**
  * Allows postponing execution of some code (command), adding it to the list via [add],
  * and performing all added commands in some time in the future via [perform]
  */
-internal class CommandList(
+private class CommandList(
     private var onNewCommand: () -> Unit
 ) {
     private val sync = createSynchronizedObject()
