+++ compose/ui/ui/src/skikoMain/kotlin/androidx/compose/ui/platform/Wrapper.skiko.kt
@@ -20,12 +20,11 @@ import androidx.compose.runtime.Composition
 import androidx.compose.runtime.CompositionContext
 import androidx.compose.runtime.CompositionLocalContext
 import androidx.compose.runtime.CompositionLocalProvider
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.remember
 import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.node.LayoutNode
 import androidx.compose.ui.node.Owner
-import androidx.compose.ui.window.RootNodeOwner
+import androidx.compose.ui.node.RootNodeOwner
 
 /**
  * Composes the given composable into [RootNodeOwner]
