+++ compose/ui/ui/src/skikoMain/kotlin/androidx/compose/ui/scene/BaseComposeScene.kt
@@ -35,6 +35,7 @@ import androidx.compose.ui.input.pointer.PointerEventType
 import androidx.compose.ui.input.pointer.PointerInputEvent
 import androidx.compose.ui.input.pointer.PointerKeyboardModifiers
 import androidx.compose.ui.input.pointer.PointerType
+import androidx.compose.ui.node.SnapshotInvalidationTracker
 import androidx.compose.ui.platform.GlobalSnapshotManager
 import kotlin.coroutines.CoroutineContext
 import kotlin.jvm.Volatile
